+++ R/ig_png.R
@@ -1,33 +1,26 @@
 #' Convert z-scores/centiles to values in the INTERGROWTH-21<sup>st</sup>
 #' Postnatal Growth Standards for preterm infants
 #'
-#' @param x Numeric vector of length one or more with x values. Elements of `x`
-#'   or its standard-specific equivalents (`pma_weeks`, `length_cm`) should have
-#'   specific units and be between certain values depending on the standard in
-#'   use (defined by `acronym`). These are:
+#' @param x,pma_weeks,length_cm Numeric vector of length one or more with x
+#'   values. Elements of `x` or its standard-specific equivalents (`pma_weeks`,
+#'   `length_cm`) should have specific units and be between certain values
+#'   depending on the standard in use (defined by `acronym`). These are:
 #'   * Between 27 and 64 weeks for `"wfa"`, `"lfa"`, and `"hcfa"`.
-#'   * Between 35 and 65 days for `"wfl"`.
+#'   * Between 35 and 65 cm for `"wfl"`.
 #'
 #'   By default, gigs will replace out-of-bounds elements in `x` with `NA` and
-#'   warn you. This behaviour can be customised using the functions in
-#'   [gigs_options].
-#' @param acronym Character vector of length one or more denoting the
-#'   INTERGROWTH-21<sup>st</sup> Postnatal Growth standard(s) in use. Each
-#'   element should be one of:
+#'   warn you. You can customise this behaviour using the [GIGS package-level
+#'   options][gigs_options].
+#' @param acronym Single-length character variable denoting the
+#'   INTERGROWTH-21<sup>st</sup> Postnatal Growth standard(s) in use. Should be
+#'   one of:
 #'   * `"wfa"` (weight-for-age)
 #'   * `"lfa"` (length-for-age)
 #'   * `"hcfa"` (head circumference-for-age)
 #'   * `"wfl"` (weight-for-length)
 #'
-#'   This argument is case-sensitive. By default, gigs will replace elements in
-#'   `acronym` which are not one of the above values with `NA` and warn you.
-#'   This behaviour can be customised using the functions in [gigs_options]. If
-#'   all elements in `acronym` are not one of the above values, gigs will throw
-#'   an error.
-#' @param pma_weeks,length_cm Numeric vector of length one or more with
-#'   standard-specific `x` variables. See the documentation for `x` for
-#'   information the acceptable bounds of these variables, and on how
-#'   out-of-bounds elements will be handled.
+#'   This argument is case-sensitive. If `acronym` is not one of the values
+#'   listed above the function will throw an error.
 #' @srrstats {G2.3b} Explicit reference to `acronym` case-sensitivity.
 #' @inherit shared_roxygen_params params note
 #' @inherit shared_zscore2value_returns return
@@ -377,11 +370,9 @@ ig_png_wfl_value2centile <- function(weight_kg, length_cm, sex) {
 #' @noRd
 ig_png_z2v_internal <- function(z, x, sex, acronym) {
   png_coeffs <- ig_png_equations(x = x, sex = sex, acronym = acronym)
-  with(png_coeffs,
-       ifelse(test = is_logarithmic,
-              yes = exp(mu_sigma_z2y(z = z, mu = mu, sigma = sigma)),
-              no = mu_sigma_z2y(z = z, mu = mu, sigma = sigma))
-  )
+  is_logarithmic <- acronym == "wfa" || acronym == "lfa"
+  y <- with(png_coeffs, mu_sigma_z2y(z = z, mu = mu, sigma = sigma))
+  if (is_logarithmic) exp(y) else y
 }
 
 #' Convert values to z-scores in the INTERGROWTH-21<sup>st</sup> Postnatal
@@ -391,12 +382,11 @@ ig_png_z2v_internal <- function(z, x, sex, acronym) {
 #' @noRd
 ig_png_v2z_internal <- function(y, x, sex, acronym) {
   png_coeffs <- ig_png_equations(x = x, sex = sex, acronym = acronym)
-  with(png_coeffs,
-       mu_sigma_y2z(
-         y = ifelse(test = is_logarithmic, yes = log(y), no = y),
-         mu = mu, sigma = sigma
-       )
-  )
+  is_logarithmic <- acronym == "wfa" || acronym == "lfa"
+  if (is_logarithmic) {
+    y <- log(y)
+  }
+  with(png_coeffs, mu_sigma_y2z(y = y, mu = mu, sigma = sigma))
 }
 
 #' INTERGROWTH-21<sup>st</sup> equations for postnatal size for age in preterm
@@ -405,17 +395,19 @@ ig_png_v2z_internal <- function(y, x, sex, acronym) {
 #' Estimates median and standard deviation for different measures of postnatal
 #' growth in preterm infants.
 #'
-#' @param sex Character vector of length one or more with sex(es), either `"M"`
-#'   (male) or `"F"` (female). This argument is case-sensitive.
-#' @param pma_weeks Numeric vector with length equal to `sex`, with
+#' @param x Numeric vector with length equal to `sex`, with
 #'   post-menstrual age(s) in exact weeks. Elements not between `27` and `64`
 #'   will return invalid results.
-#' @param acronym Character vector of length one or more with acronym(s)
-#'   denoting an INTERGROWTH-21<sup>st</sup> Postnatal Growth standard. Should
-#'   be one of `"wfa"` (weight-for-age), `"lfa"` (length-for-age), `"hcfa"`
-#'   (head circumference-for-age), or `"wfl"` (weight-for-length).
-#' @return A data frame with median(s) and standard deviation(s) for each
-#'   `age`/`sex`/`acronym` combination provided to the function.
+#' @param sex Character vector of length one or more with sex(es), either `"M"`
+#'   (male) or `"F"` (female). This argument is case-sensitive.
+#' @param acronym A single-length character vector denoting an
+#'   INTERGROWTH-21<sup>st</sup> Postnatal Growth standard. Should be one of
+#'   `"wfa"` (weight-for-age), `"lfa"` (length-for-age), `"hcfa"` (head
+#'   circumference-for-age), or `"wfl"` (weight-for-length).
+#' @return A named list of two numeric vectors, where each element corresponds
+#'   to the element-wise combination of `x` and `sex`:
+#'   * `mu` - mean value at `x` for this `sex`
+#'   * `sigma` - standard deviation at `x` for this `sex`
 #' @note The weight-for-age and length-for-age standards are logarithmic, so
 #'   require slightly different treatment to use in z-score conversions. In
 #'   contrast, head circumference for gestational age returns the median and
@@ -430,72 +422,48 @@ ig_png_v2z_internal <- function(y, x, sex, acronym) {
 #' @rdname ig_png_equations
 #' @noRd
 ig_png_equations <- function(x, sex, acronym) {
-  wfa_log_mu <- function(pma_weeks, sex) {
-    2.591277 - (0.01155 * (pma_weeks^0.5)) - (2201.705 * (pma_weeks^-2)) +
-      (0.0911639 * sex)
-  }
-  lfa_log_mu <- function(pma_weeks, sex) {
-    4.136244 - (547.0018 * (pma_weeks^-2)) + 0.0026066 * pma_weeks +
-      0.0314961 * sex
-  }
-  hcfa_mu <- function(pma_weeks, sex) {
-    55.53617 - (852.0059 * (pma_weeks^-1)) + 0.7957903 * sex
-  }
-  wfl_mu <- function(length_cm, sex) {
-    ifelse(
-      sex == "M",
-      yes = 13.98383 + 203.5677 * (length_cm / 10)^-2 - 291.114 *
-        ((length_cm / 10)^ -2 * log(length_cm/10)),
-      no = ifelse(
-        sex == "F",
-        yes = 50.32492 + 140.8019 * (length_cm / 10)^-1 - 167.906 *
-          (length_cm / 10)^-0.5,
-        no = NA_real_))
-  }
-  wfa_sigma <- function(pma_weeks) {
-    0.1470258 + 505.92394 / pma_weeks^2 -
-      140.0576 / (pma_weeks^2) * log(pma_weeks)
-  }
-  lfa_sigma <- function(pma_weeks) {
-    0.050489 + (310.44761 * (pma_weeks^-2)) -
-      (90.0742 * (pma_weeks^-2)) * log(pma_weeks)
-  }
-  hcfa_sigma <- function(pma_weeks) {
-    3.0582292 + (3910.05 * (pma_weeks^-2)) - 180.5625 * pma_weeks^-1
-  }
-  wfl_sigma <- function(length_cm, sex) {
-    ifelse(
-      sex == "M",
-      yes = exp(-1.830098 + 0.0049708 * (length_cm / 10)^3),
-      no = ifelse(
-        sex == "F",
-        yes = 0.2195888 - 0.0046046 * (length_cm / 10)^3 + 0.0033017 *
-        (length_cm / 10)^3 * log(length_cm / 10),
-        no = NA_real_))
-  }
-  out_df <- data.frame(x = x, sex = sex, acronym = acronym)
-  sex_as_numeric <- ifelse(sex == "M", yes = 1, no = 0)
-  out_df[["mu"]] <- ifelse(
-    acronym == "wfa",
-    yes = wfa_log_mu(out_df[["x"]], sex_as_numeric),
-    no = ifelse(acronym == "lfa",
-                yes = lfa_log_mu(out_df[["x"]], sex_as_numeric),
-                no = ifelse(acronym == "hcfa",
-                            hcfa_mu(out_df[["x"]], sex_as_numeric),
-                            wfl_mu(out_df[["x"]], sex))))
-  out_df[["sigma"]] <- ifelse(
-    acronym == "wfa",
-    yes = wfa_sigma(out_df[["x"]]),
-    no = ifelse(acronym == "lfa",
-                yes = lfa_sigma(out_df[["x"]]),
-                no = ifelse(acronym == "hcfa",
-                            yes = hcfa_sigma(out_df[["x"]]),
-                            no = wfl_sigma(out_df[["x"]], sex)))
-  )
-  out_df[["is_logarithmic"]] <- acronym %in% c("wfa", "lfa")
-  invalid_params <- !stats::complete.cases(out_df)
-  out_df[["mu"]][invalid_params] <- NA
-  out_df
+  sex_as_numeric <- sex == "M"
+  switch(acronym,
+         wfa = list(
+           mu = 2.591277 - 0.01155 * x^0.5 - 2201.705 * x^(-2) +
+             0.0911639 * sex_as_numeric,
+           sigma = 0.1470258 + 505.92394 / x^2 - 140.0576 / x^2 * log(x)
+         ),
+         lfa = list(
+           mu = 4.136244 - 547.0018 * x^(-2) + 0.0026066 * x +
+             0.0314961 * sex_as_numeric,
+           sigma = 0.050489 + 310.44761 * x^(-2) - 90.0742 * x^(-2) * log(x)
+         ),
+         hcfa = list(
+           mu = 55.53617 - 852.0059 * x^(-1) + 0.7957903 * sex_as_numeric,
+           sigma = 3.0582292 + 3910.05 * x^(-2) - 180.5625 * x^(-1)
+         ),
+         wfl = {
+           length_dm <- x / 10
+           log_length <- log(length_dm)
+           length_cubed <- length_dm^3
+           list(
+             mu = ifelse(
+               sex == "M",
+               yes = 13.98383 + 203.5677 * length_dm^(-2) - 291.114 *
+                 length_dm^(-2) * log_length,
+               no = ifelse(
+                 sex == "F",
+                 yes = 50.32492 + 140.8019 * length_dm^(-1) - 167.906 *
+                   length_dm^(-0.5),
+                 no = NA_real_
+               )),
+             sigma = ifelse(
+               sex == "M",
+               yes = exp(-1.830098 + 0.0049708 * length_cubed),
+               no = ifelse(
+                 sex == "F",
+                 yes = 0.2195888 - 0.0046046 * length_cubed + 0.0033017 *
+                   length_cubed * log_length,
+                 no = NA_real_
+               ))
+           )
+         })
 }
 
 # SRR tags ---------------------------------------------------------------------