+++ R/ig_vpns.R
@@ -8,11 +8,11 @@
 #'   in days. Elements not between `168` and `230` will return invalid output.
 #' @param sex Character vector of length one or more with sex(es), either `"M"`
 #'   (male) or `"F"` (female). This argument is case-sensitive.
-#' @param acronym Character vector of length one or more with acronym(s)
-#'   denoting the INTERGROWTH-21<sup>st</sup> VPNS standard to use. Each element
-#'   should be one of `"wfga"`, `"lfga"`, or `"hcfga"`.
+#' @param acronym A single string denoting the INTERGROWTH-21<sup>st</sup> VPNS
+#'   standard to use. Must be one of `"wfga"`, `"lfga"`, or `"hcfga"` and is
+#'   case-sensitive.
 #' @return A data frame with mean and standard deviation values for each
-#' provided combination of sex, gestational age, and acronym.
+#'   elementwise combination of `sex` and `gest_days`.
 #' @note This function returns the **natural log** of the median and standard
 #' deviations for weight (kg) for gestational age. In contrast, the medians and
 #' standard deviations for length and head circumference for gestational age
@@ -27,42 +27,23 @@
 #' @importFrom stats complete.cases
 #' @noRd
 ig_vpns_equations <- function(gest_days, sex, acronym) {
-  gest_days[gest_days >= 231] <- NA
-  wfga_logmu <- function(ga_weeks, sex) {
-    -7.00303 + 1.325911 * ga_weeks^0.5 + 0.0571937 * sex
-  }
-  lfga_mu <- function(ga_weeks, sex) {
-    1.307633 + 1.270022 * ga_weeks +  0.4263885 * sex
-  }
-  hcfga_mu <- function(ga_weeks, sex) {
-    0.7866522 + 0.887638 * ga_weeks + 0.2513385 * sex
-  }
-  wfga_sigma <- sqrt(x = 0.0373218)
-  lfga_sigma <- sqrt(x = 6.757543)
-  hcfga_sigma <- sqrt(x = 2.433481)
-
-  incomplete <- !complete.cases(gest_days, sex, acronym)
-  gest_days[incomplete] <- NA
-  sex[incomplete] <- NA
-  acronym[incomplete] <- NA
-
+  gest_days[gest_days >= 231] <- NA_real_
+  ga_weeks <- gest_days / 7
   sex_as_numeric <- ifelse(sex == "M", yes = 1, no = 0)
-  gest_weeks <- gest_days / 7
-  out_len <- length(gest_weeks)
-  out <- list(mu = numeric(length = out_len),
-              sigma = numeric(length = out_len))
-  out[["mu"]] <- ifelse(acronym == "wfga",
-                        yes = wfga_logmu(gest_weeks, sex_as_numeric),
-                        no = ifelse(acronym == "lfga",
-                                    yes = lfga_mu(gest_weeks, sex_as_numeric),
-                                    no = hcfga_mu(gest_weeks, sex_as_numeric)))
-  out[["sigma"]] <- ifelse(acronym == "wfga",
-                           yes = wfga_sigma,
-                           no = ifelse(acronym == "lfga",
-                                       yes = lfga_sigma,
-                                       no = hcfga_sigma))
-  out[["logarithmic"]] <- acronym == "wfga"
-  out
+  switch(acronym,
+    wfga = list(
+      mu = -7.00303 + 1.325911 * ga_weeks^0.5 + 0.0571937 * sex_as_numeric,
+      sigma = sqrt(x = 0.0373218)
+    ),
+    lfga = list(
+      mu = 1.307633 + 1.270022 * ga_weeks +  0.4263885 * sex_as_numeric,
+      sigma = sqrt(x = 6.757543)
+    ),
+    hcfga = list(
+      mu = 0.7866522 + 0.887638 * ga_weeks + 0.2513385 * sex_as_numeric,
+      sigma = sqrt(x = 2.433481)
+    )
+  )
 }
 
 #' Convert z-scores to values in the INTERGROWTH-21<sup>st</sup> Newborn Size
@@ -74,11 +55,11 @@ ig_vpns_equations <- function(gest_days, sex, acronym) {
 #'   in days. Elements not between `168` and `230` will return invalid output.
 #' @param sex Character vector of length one or more with sex(es), either `"M"`
 #'   (male) or `"F"` (female). This argument is case-sensitive.
-#' @param acronym Character vector of length one or more with acronym(s)
-#'   denoting the INTERGROWTH-21<sup>st</sup> VPNS standard to use. Each element
-#'   should be one of `"wfga"`, `"lfga"`, or `"hcfga"`.
+#' @param acronym Single-length character variable with an acronym
+#'   denoting the INTERGROWTH-21<sup>st</sup> VPNS standard to use. Should be
+#'   one of `"wfga"`, `"lfga"`, or `"hcfga"`.
 #' @returns Numeric vector the same length as `z` with expected measurements for
-#'   each element of `z`, `gest_days`, `sex`, and `acronym` provided to the
+#'   each element of `z`, `gest_days`, and `sex` provided to the
 #'   function.
 #' @references
 #' Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
@@ -89,14 +70,8 @@ ig_vpns_zscore2value <- function(z, gest_days, sex, acronym) {
   mu_sigma <- ig_vpns_equations(gest_days = gest_days,
                                 sex = sex,
                                 acronym = acronym)
-  ifelse(test = mu_sigma[["logarithmic"]],
-         yes = exp(mu_sigma_z2y(z = z,
-                                mu = mu_sigma[["mu"]],
-                                sigma = mu_sigma[["sigma"]])),
-         no = mu_sigma_z2y(z = z,
-                           mu = mu_sigma[["mu"]],
-                           sigma = mu_sigma[["sigma"]])
-  )
+  y <- with(mu_sigma, mu_sigma_z2y(z = z, mu = mu, sigma = sigma))
+  if (acronym == "wfga") exp(y) else y
 }
 
 #' Convert values to z-scores in the INTERGROWTH-21<sup>st</sup> Newborn Size
@@ -107,11 +82,11 @@ ig_vpns_zscore2value <- function(z, gest_days, sex, acronym) {
 #'   in days. Elements not between `168` and `230` will return invalid output.
 #' @param sex Character vector of length one or more with sex(es), either `"M"`
 #'   (male) or `"F"` (female). This argument is case-sensitive.
-#' @param acronym Character vector of length one or more with acronym(s)
-#'   denoting the INTERGROWTH-21<sup>st</sup> VPNS standard to use. Each element
-#'   should be one of `"wfga"`, `"lfga"`, or `"hcfga"`.
+#' @param acronym Single-length character variable with an acronym
+#'   denoting the INTERGROWTH-21<sup>st</sup> VPNS standard to use. Should be
+#'   one of `"wfga"`, `"lfga"`, or `"hcfga"`.
 #' @returns Numeric vector the same length as `y` with z-scores for each element
-#'  of `y`, `gest_days`, `sex`, and `acronym` provided to the function.
+#'  of `y`, `gest_days`, and `sex` provided to the function.
 #' @references
 #' Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
 #' **INTERGROWTH-21st very preterm size at birth reference charts.** *Lancet*
@@ -122,15 +97,10 @@ ig_vpns_value2zscore <- function(y, gest_days, sex, acronym) {
   mu_sigma <- ig_vpns_equations(gest_days = gest_days,
                                 sex = sex,
                                 acronym = acronym)
-  ifelse(
-      test = mu_sigma[["logarithmic"]],
-      yes = mu_sigma_y2z(y = log(y),
-                         mu = mu_sigma[["mu"]],
-                         sigma = mu_sigma[["sigma"]]),
-      no = mu_sigma_y2z(y = y,
-                        mu = mu_sigma[["mu"]],
-                        sigma = mu_sigma[["sigma"]])
-  )
+  if (acronym == "wfga") {
+    y <- log(y)
+  }
+  with(mu_sigma, mu_sigma_y2z(y = y, mu = mu, sigma = sigma))
 }
 
 # SRR tags ---------------------------------------------------------------------